Use a larger scalar (64 bits on Linux) to store the total number of sites crossed including those for which a sample was not actually taken.  32 bits wraps around too quickly.

+++ mining/build-tables.sql
@@ -9,7 +9,7 @@ CREATE TABLE sessions (
 
 CREATE TABLE sites (
   session int NOT NULL CHECK (session > 0) REFERENCES sessions MATCH FULL,
-  sample int NOT NULL CHECK (sample > 0),
+  sample bigint NOT NULL CHECK (sample > 0),
   file text NOT NULL CHECK (file <> ''),
   line int NOT NULL CHECK (line > 0),
